pyproject.toml
[project]
name = "playwright-pytest"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "hemanth-kumar-j", email = "hemanth@nitimis.com" }
]
dependencies = [
    "pytest-playwright>=0.7.0",
    "pytest-xdist>=3.6.1",
    "pytest-order>=1.3.0",
    "pytest-html>=4.1.1",
    "black>=25.1.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/playwright_pytest"]
